requirements.txt
#pip install -r requirements.txt

#lines preceded by # are comments
#used to parse arguments that are passed to a script via command line
argparse~=1.4.0

#used by wordclouds.py (also uses stanza; see below)
wordcloud~=1.8.1
pillow~=8.1.0

#used by nominalization
pywsd~=1.2.4

#used by stats_visuals_util.text_readability
textstat~=0.7.0

#used by split files
mlconjug~=3.4.0

#used by pdf_converter
pdfminer.six

#used by docx_converter
python-docx

# used by YAGO
SPARQLWrapper
stanza
fuzzywuzzy

#used by spell checkers
textblob
autocorrect~=2.3.0
pyspellchecker
#language detection packages
#spacy installed below
langdetect
spacy_langdetect
langid

#used by Stanford_CoreNLP.py and other functions
unicode
unidecode
pycorenlp==0.3.0
stanfordcorenlp~=3.9.1.1
conllu
nltk~=3.5
# used by wordcloud script (the new StanfordCoreNLP Python approach)
stanza

#used by SVO
#stanfordnlp is the new stanford coreNLP Python version of their software released in 2020
torch===1.6.0
torchvision===0.7.0 -f https://download.pytorch.org/whl/torch_stable.html
#stanfordnlp

# VADER (sentimentt analysis)
twython

#used by Excel_util.py for visualzing Excel charts
openpyxl~=3.0.6

#used by GIS_main.py for generating Google Earth KML files, wordclouds_main
geopy~=2.1.0
simplekml~=1.3.5
googlemaps~=4.4.2
# tkcolorpicker requires tkinter and pillow to be installed (https://libraries.io/pypi/tkcolorpicker)
tkcolorpicker~=2.1.3

#used by Gensim, topic modeling script
gensim~=3.8.3
IPython # THIS SHOULD BE AUTOMATICALLY INSTALLED BY ANACONDA; but... not!
spacy~=2.3.5
pyLDAvis~=2.1.2
# pprint #THIS SHOULD BE AUTOMATICALLY INSTALLED BY ANACONDA
# matplotlib #THIS SHOULD BE AUTOMATICALLY INSTALLED BY ANACONDA


pandas~=1.2.1
future~=0.18.2
numpy~=1.19.2
lxml~=4.6.2
python-dateutil~=2.8.1
matplotlib~=3.3.3
scipy~=1.6.0
chardet~=4.0.0
tqdm~=4.55.1
scikit-learn~=0.24.1